+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/DocumentMapping/LSPDocumentMappingProvider.cs
@@ -5,9 +5,8 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
-using Microsoft.VisualStudio.LanguageServer.Protocol;
 
-namespace Microsoft.VisualStudio.LanguageServerClient.Razor.HtmlCSharp;
+namespace Microsoft.VisualStudio.LanguageServerClient.Razor.DocumentMapping;
 
 internal abstract class LSPDocumentMappingProvider
 {
