+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/DocumentMapping/RazorMapToDocumentRangesParams.cs
@@ -7,7 +7,7 @@ using System.Runtime.Serialization;
 using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
 using Microsoft.Extensions.Internal;
 
-namespace Microsoft.VisualStudio.LanguageServerClient.Razor.HtmlCSharp;
+namespace Microsoft.VisualStudio.LanguageServerClient.Razor.DocumentMapping;
 
 // Note: This type should be kept in sync with the one in Razor.LanguageServer assembly.
 [DataContract]
